+++ gui/src/main/java/engine/gui/graphics/control/CheckBoxRenderer.java
@@ -16,13 +16,13 @@ public class CheckBoxRenderer extends NodeRenderer<CheckBox> {
             float width;
             float height;
             if (node.state().get() == null) {
-                width = node.getWidth() - insets.getRight() - insets.getLeft();
-                height = node.getWidth() - insets.getTop() - insets.getBottom();
+                width = (float) (node.getWidth() - insets.getRight() - insets.getLeft());
+                height = (float) (node.getWidth() - insets.getTop() - insets.getBottom());
                 graphics.setColor(node.contentColor().get());
-                graphics.fillRect(insets.getLeft(), insets.getTop(), width, height);
+                graphics.fillRect((float) insets.getLeft(), (float) insets.getTop(), width, height);
             } else {
-                width = node.getWidth();
-                height = node.getWidth();
+                width = (float) node.getWidth();
+                height = (float) node.getWidth();
                 float vc = (float) Math.sqrt(Math.pow(node.getBorder().getInsets().getLeft(), 2) / 2);
                 graphics.fillQuad(new Vector2f(vc, 0), new Vector2f(width, height - vc), new Vector2f(width - vc, height), new Vector2f(0, vc));
                 graphics.fillQuad(new Vector2f(0, height - vc), new Vector2f(width - vc, 0), new Vector2f(width, vc), new Vector2f(vc, height));
