+++ gui/src/main/java/engine/gui/graphics/control/RadioButtonRenderer.java
@@ -12,10 +12,10 @@ public class RadioButtonRenderer extends NodeRenderer<RadioButton> {
         graphics.drawBorder(node.getBorder(), node);
         if (node.selected().get()) {
             Insets insets = node.getPadding();
-            float width = node.getWidth() - insets.getRight() - insets.getLeft();
-            float height = node.getWidth() - insets.getTop() - insets.getBottom();
+            float width = (float) (node.getWidth() - insets.getRight() - insets.getLeft());
+            float height = (float) (node.getWidth() - insets.getTop() - insets.getBottom());
             graphics.setColor(node.contentColor().get());
-            graphics.fillRect(insets.getLeft(), insets.getTop(), width, height);
+            graphics.fillRect((float) insets.getLeft(), (float) insets.getTop(), width, height);
         }
     }
 }
